# [sdk-metrics] Tags perf improvements