--- a/src/OpenTelemetry/Metrics/AggregatorStore.cs
+++ b/src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -37,9 +37,7 @@ internal sealed class AggregatorStore
 
     // This holds the reclaimed MetricPoints that are available for reuse.
     private readonly Queue<int>? availableMetricPoints;
-
-    private readonly ConcurrentDictionary<Tags, int> tagsToMetricPointIndexDictionary =
-        new();
+    private readonly ConcurrentDictionary<Tags, int>? tagsToMetricPointIndexDictionary;
 
     private readonly string name;
     private readonly string metricPointCapHitMessage;
@@ -122,15 +120,16 @@ internal sealed class AggregatorStore
 
         this.OutputDeltaWithUnusedMetricPointReclaimEnabled = shouldReclaimUnusedMetricPoints && this.OutputDelta;
 
+        // Note: We expect at the most (maxMetricPoints -
+        // reservedMetricPointsCount) * 2 entries (one for sorted and one for
+        // unsorted input)
+        var maximumNumberOfMetricPoints = (cardinalityLimit - reservedMetricPointsCount) * 2;
+
         if (this.OutputDeltaWithUnusedMetricPointReclaimEnabled)
         {
             this.availableMetricPoints = new Queue<int>(cardinalityLimit - reservedMetricPointsCount);
 
-            // There is no overload which only takes capacity as the parameter
-            // Using the DefaultConcurrencyLevel defined in the ConcurrentDictionary class: https://github.com/dotnet/runtime/blob/v7.0.5/src/libraries/System.Collections.Concurrent/src/System/Collections/Concurrent/ConcurrentDictionary.cs#L2020
-            // We expect at the most (maxMetricPoints - reservedMetricPointsCount) * 2 entries- one for sorted and one for unsorted input
-            this.TagsToMetricPointIndexDictionaryDelta =
-                new ConcurrentDictionary<Tags, LookupData>(concurrencyLevel: Environment.ProcessorCount, capacity: (cardinalityLimit - reservedMetricPointsCount) * 2);
+            this.TagsToMetricPointIndexDictionaryDelta = new(concurrencyLevel: Environment.ProcessorCount, capacity: maximumNumberOfMetricPoints);
 
             // Add all the indices except for the reserved ones to the queue so that threads have
             // readily available access to these MetricPoints for their use.
@@ -143,6 +142,7 @@ internal sealed class AggregatorStore
         }
         else
         {
+            this.tagsToMetricPointIndexDictionary = new(concurrencyLevel: Environment.ProcessorCount, capacity: maximumNumberOfMetricPoints);
             this.lookupAggregatorStore = this.LookupAggregatorStore;
         }
     }
@@ -457,9 +457,11 @@ internal sealed class AggregatorStore
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private int LookupAggregatorStore(KeyValuePair<string, object?>[] tagKeysAndValues, int length)
     {
+        Debug.Assert(this.tagsToMetricPointIndexDictionary != null, "this.tagsToMetricPointIndexDictionary was null");
+
         var givenTags = new Tags(tagKeysAndValues);
 
-        if (!this.tagsToMetricPointIndexDictionary.TryGetValue(givenTags, out var aggregatorIndex))
+        if (!this.tagsToMetricPointIndexDictionary!.TryGetValue(givenTags, out var aggregatorIndex))
         {
             if (length > 1)
             {