@cijothomas 2024-03-18
> @cijothomas Added

Given we are ultra optimized for <=3 tags, and reasonably optimized for 3-7 tags, it'd be nicer to use a tagCount<7 in benchmarks shared, to see more practical gains.

Also, if feasible, can you show the existing benchmarks before/after, so as to get a picture of how much this affects overall perf. (hash calc and equals() is significant chunk of overall metrics cost in hot path, so I expect this to be significant. But curios if it is significant enough to show up in normal benchmark runs)