@CodeBlanch 2024-03-19
@cijothomas I updated the description to show benchmarks for different number of tags. There aren't any optimizations in `struct Tags` based on length, the gains should be pretty linear based on the number of tags. It might be worth looking at if we should special case a length of 1?

Here is a run of `MetricsBenchmarks`:

Before:

| Method                                | AggregationTemporality | Mean      | Error    | StdDev   |
|-------------------------------------- |----------------------- |----------:|---------:|---------:|
| CounterHotPath                        | Cumulative             |  13.01 ns | 0.146 ns | 0.136 ns |
| CounterWith1LabelsHotPath             | Cumulative             |  53.83 ns | 0.330 ns | 0.293 ns |
| CounterWith2LabelsHotPath             | Cumulative             |  72.84 ns | 0.783 ns | 0.695 ns |
| CounterWith3LabelsHotPath             | Cumulative             |  99.68 ns | 0.652 ns | 0.578 ns |
| CounterWith4LabelsHotPath             | Cumulative             | 125.77 ns | 1.449 ns | 1.356 ns |
| CounterWith5LabelsHotPath             | Cumulative             | 144.82 ns | 1.459 ns | 1.364 ns |
| CounterWith6LabelsHotPath             | Cumulative             | 164.60 ns | 1.171 ns | 0.978 ns |
| CounterWith7LabelsHotPath             | Cumulative             | 184.49 ns | 2.803 ns | 2.485 ns |
| CounterWith1LabelsHotPathUsingTagList | Cumulative             |  64.65 ns | 1.312 ns | 1.227 ns |
| CounterWith2LabelsHotPathUsingTagList | Cumulative             |  87.66 ns | 1.096 ns | 0.972 ns |
| CounterWith3LabelsHotPathUsingTagList | Cumulative             | 114.65 ns | 1.031 ns | 0.964 ns |
| CounterWith4LabelsHotPathUsingTagList | Cumulative             | 135.71 ns | 1.352 ns | 1.265 ns |
| CounterWith5LabelsHotPathUsingTagList | Cumulative             | 156.48 ns | 1.008 ns | 0.943 ns |
| CounterWith6LabelsHotPathUsingTagList | Cumulative             | 180.64 ns | 1.736 ns | 1.624 ns |
| CounterWith7LabelsHotPathUsingTagList | Cumulative             | 202.70 ns | 2.262 ns | 2.116 ns |
| CounterWith8LabelsHotPathUsingTagList | Cumulative             | 228.39 ns | 2.670 ns | 2.498 ns |
| CounterWith9LabelsHotPathUsingTagList | Cumulative             | 294.53 ns | 2.080 ns | 1.844 ns |
| CounterHotPath                        | Delta                  |  12.62 ns | 0.039 ns | 0.034 ns |
| CounterWith1LabelsHotPath             | Delta                  |  52.23 ns | 0.441 ns | 0.413 ns |
| CounterWith2LabelsHotPath             | Delta                  |  73.28 ns | 0.727 ns | 0.680 ns |
| CounterWith3LabelsHotPath             | Delta                  |  97.32 ns | 1.266 ns | 1.184 ns |
| CounterWith4LabelsHotPath             | Delta                  | 124.98 ns | 2.041 ns | 1.909 ns |
| CounterWith5LabelsHotPath             | Delta                  | 150.81 ns | 2.569 ns | 3.685 ns |
| CounterWith6LabelsHotPath             | Delta                  | 170.89 ns | 1.589 ns | 1.408 ns |
| CounterWith7LabelsHotPath             | Delta                  | 185.56 ns | 1.646 ns | 1.540 ns |
| CounterWith1LabelsHotPathUsingTagList | Delta                  |  62.80 ns | 0.481 ns | 0.426 ns |
| CounterWith2LabelsHotPathUsingTagList | Delta                  |  89.48 ns | 0.794 ns | 0.743 ns |
| CounterWith3LabelsHotPathUsingTagList | Delta                  | 114.09 ns | 1.596 ns | 1.492 ns |
| CounterWith4LabelsHotPathUsingTagList | Delta                  | 135.33 ns | 2.359 ns | 2.206 ns |
| CounterWith5LabelsHotPathUsingTagList | Delta                  | 155.94 ns | 0.978 ns | 0.915 ns |
| CounterWith6LabelsHotPathUsingTagList | Delta                  | 180.58 ns | 0.955 ns | 0.894 ns |
| CounterWith7LabelsHotPathUsingTagList | Delta                  | 200.10 ns | 1.079 ns | 1.009 ns |
| CounterWith8LabelsHotPathUsingTagList | Delta                  | 228.79 ns | 2.795 ns | 2.615 ns |
| CounterWith9LabelsHotPathUsingTagList | Delta                  | 297.95 ns | 4.772 ns | 4.464 ns |

After:

| Method                                | AggregationTemporality | Mean      | Error    | StdDev   |
|-------------------------------------- |----------------------- |----------:|---------:|---------:|
| CounterHotPath                        | Cumulative             |  12.61 ns | 0.271 ns | 0.371 ns |
| CounterWith1LabelsHotPath             | Cumulative             |  46.16 ns | 0.480 ns | 0.449 ns |
| CounterWith2LabelsHotPath             | Cumulative             |  62.84 ns | 1.123 ns | 1.051 ns |
| CounterWith3LabelsHotPath             | Cumulative             |  92.78 ns | 1.874 ns | 4.033 ns |
| CounterWith4LabelsHotPath             | Cumulative             | 117.40 ns | 1.908 ns | 1.691 ns |
| CounterWith5LabelsHotPath             | Cumulative             | 136.79 ns | 2.619 ns | 3.671 ns |
| CounterWith6LabelsHotPath             | Cumulative             | 155.58 ns | 1.922 ns | 1.798 ns |
| CounterWith7LabelsHotPath             | Cumulative             | 171.33 ns | 2.661 ns | 2.489 ns |
| CounterWith1LabelsHotPathUsingTagList | Cumulative             |  60.23 ns | 0.851 ns | 0.796 ns |
| CounterWith2LabelsHotPathUsingTagList | Cumulative             |  79.14 ns | 1.011 ns | 0.945 ns |
| CounterWith3LabelsHotPathUsingTagList | Cumulative             | 104.30 ns | 1.196 ns | 1.119 ns |
| CounterWith4LabelsHotPathUsingTagList | Cumulative             | 122.29 ns | 1.709 ns | 1.427 ns |
| CounterWith5LabelsHotPathUsingTagList | Cumulative             | 141.39 ns | 1.343 ns | 1.256 ns |
| CounterWith6LabelsHotPathUsingTagList | Cumulative             | 163.88 ns | 1.119 ns | 1.046 ns |
| CounterWith7LabelsHotPathUsingTagList | Cumulative             | 185.83 ns | 2.857 ns | 2.672 ns |
| CounterWith8LabelsHotPathUsingTagList | Cumulative             | 209.42 ns | 1.477 ns | 1.310 ns |
| CounterWith9LabelsHotPathUsingTagList | Cumulative             | 272.17 ns | 4.786 ns | 4.477 ns |
| CounterHotPath                        | Delta                  |  12.49 ns | 0.101 ns | 0.089 ns |
| CounterWith1LabelsHotPath             | Delta                  |  45.07 ns | 0.256 ns | 0.227 ns |
| CounterWith2LabelsHotPath             | Delta                  |  63.01 ns | 0.270 ns | 0.252 ns |
| CounterWith3LabelsHotPath             | Delta                  |  86.59 ns | 1.717 ns | 1.522 ns |
| CounterWith4LabelsHotPath             | Delta                  | 116.75 ns | 2.345 ns | 4.575 ns |
| CounterWith5LabelsHotPath             | Delta                  | 131.35 ns | 2.530 ns | 2.707 ns |
| CounterWith6LabelsHotPath             | Delta                  | 150.90 ns | 2.327 ns | 1.943 ns |
| CounterWith7LabelsHotPath             | Delta                  | 173.34 ns | 3.479 ns | 5.099 ns |
| CounterWith1LabelsHotPathUsingTagList | Delta                  |  58.93 ns | 0.993 ns | 0.880 ns |
| CounterWith2LabelsHotPathUsingTagList | Delta                  |  81.29 ns | 1.625 ns | 2.577 ns |
| CounterWith3LabelsHotPathUsingTagList | Delta                  | 104.42 ns | 1.757 ns | 1.644 ns |
| CounterWith4LabelsHotPathUsingTagList | Delta                  | 126.46 ns | 2.380 ns | 2.226 ns |
| CounterWith5LabelsHotPathUsingTagList | Delta                  | 146.67 ns | 2.696 ns | 2.521 ns |
| CounterWith6LabelsHotPathUsingTagList | Delta                  | 169.62 ns | 1.926 ns | 1.801 ns |
| CounterWith7LabelsHotPathUsingTagList | Delta                  | 195.53 ns | 3.848 ns | 3.951 ns |
| CounterWith8LabelsHotPathUsingTagList | Delta                  | 205.62 ns | 2.932 ns | 2.743 ns |
| CounterWith9LabelsHotPathUsingTagList | Delta                  | 281.14 ns | 4.489 ns | 4.199 ns |